Honor request filter abort and priority semantics

+++ src/main/java/io/muserver/rest/RestHandler.java
@@ -80,6 +80,10 @@ public class RestHandler implements MuHandler {
         JaxRSRequest requestContext = new JaxRSRequest(muRequest, muResponse, new LazyAccessInputStream(muRequest), Mutils.trim(muRequest.relativePath(), "/"), securityContext, readerInterceptors, entityProviders);
         try {
             filterManagerThing.onPreMatch(requestContext);
+            if (requestContext.getAbortResponse() != null) {
+                sendResponse(0, requestContext, muResponse, acceptHeaders, emptyList(), emptyList(), JaxRSResponse.Builder.EMPTY_ANNOTATIONS, requestContext.getAbortResponse());
+                return true;
+            }
 
             Function<RequestMatcher.MatchedMethod,ResourceClass> subResourceLocator = matchedMethod -> {
                 Function<ResourceMethod, Object> onSuspended = resourceMethod -> {
@@ -121,6 +125,10 @@ public class RestHandler implements MuHandler {
             Annotation[] methodAnnotations = mm.resourceMethod.methodAnnotations;
 
             filterManagerThing.onPostMatch(requestContext);
+            if (requestContext.getAbortResponse() != null) {
+                sendResponse(0, requestContext, muResponse, acceptHeaders, produces, directlyProduces, JaxRSResponse.Builder.EMPTY_ANNOTATIONS, requestContext.getAbortResponse());
+                return true;
+            }
 
             Function<ResourceMethod, Object> suspendedParamCallback = rm -> {
                 if (muRequest.isAsync()) {
